feat: Add TextLabel color set to Assets for improved UI theming with light and dark appearances
